Fix unit tests (#28)
* Remove console.* and replace with logger where needed

* Add GitHub workflow for unit tests

* Fix unit tests

Todo: break up media player
