+++ src/command/PlayAOEFileCommand.ts
@@ -1,23 +1,24 @@
 import { IMediaFilePlayer } from './../media/MediaFilePlayer';
-import { createAudioPlayer, createAudioResource, joinVoiceChannel } from '@discordjs/voice';
 import { SuccessfulParsedMessage } from 'discord-command-parser';
 import { Message } from 'discord.js';
 import fs from 'fs';
 import { ICommand } from './ICommand';
 import { isInteger, leftPad } from '../helpers/helpers';
+import { Logger } from 'winston';
 
 const AOE_SOUND_DIRECTORY = `${process.cwd()}\\data\\sounds\\age_taunts`;
 
 export class PlayAOEFileCommand implements ICommand {
     public constructor(
-        private mediaPlayer: IMediaFilePlayer
+        private mediaPlayer: IMediaFilePlayer,
+        private logger: Logger
     ) { }
     
     async execute(cmd: SuccessfulParsedMessage<Message>, msg: Message): Promise<void> {
         let { body } = cmd;
 
         if (!body || !isInteger(body)) {
-            console.error('No body given');
+            this.logger.error('No body given');
         }
 
         body = leftPad(Number(body), 2);
@@ -27,13 +28,13 @@ export class PlayAOEFileCommand implements ICommand {
         try {
             fileNames = await fs.promises.readdir(AOE_SOUND_DIRECTORY);
         } catch (error) {
-            console.error({ errorWhileFindingAgeTaunts: error });
+            this.logger.error({ errorWhileFindingAgeTaunts: error });
         }
 
         const fileName = fileNames.find((fileName: string) => fileName.startsWith(body));
 
         if (!fileName) {
-            console.error('File not found: ' + body);
+            this.logger.error('File not found: ' + body);
             return;
         }
